verbose(1)

+++ proactive_conflict_check.sh
@@ -42,7 +42,9 @@ echo "----------" $temporary_branch "----------"
 echo "----------" $cleartax_triggering_ref "----------"
 
 #remove any existing error log file
-rm $error_log_file || die "Failed to delete error log file"
+if [ -f $error_log_file ]; then
+	rm $error_log_file || die "Failed to delete error log file"
+fi
 
 cd "$base_production_code_dir" || die "Failed to go to directory $base_production_code_dir"
